creation of signUp LogIn page and userInputs + tests
