--- app/src/androidTest/java/ch/epfl/sweng/hostme/UserCreation4Test.java
@@ -1,46 +1,0 @@
-package ch.epfl.sweng.hostme;
-
-import static androidx.test.core.app.ActivityScenario.launch;
-import static androidx.test.core.app.ApplicationProvider.getApplicationContext;
-import static androidx.test.espresso.Espresso.onView;
-import static androidx.test.espresso.action.ViewActions.clearText;
-import static androidx.test.espresso.action.ViewActions.click;
-import static androidx.test.espresso.action.ViewActions.closeSoftKeyboard;
-import static androidx.test.espresso.action.ViewActions.typeText;
-import static androidx.test.espresso.assertion.ViewAssertions.matches;
-import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
-import static androidx.test.espresso.matcher.ViewMatchers.withId;
-import static androidx.test.espresso.matcher.ViewMatchers.withText;
-
-import static org.junit.Assert.assertTrue;
-
-import android.content.Intent;
-
-import androidx.test.core.app.ActivityScenario;
-import androidx.test.ext.junit.rules.ActivityScenarioRule;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
-
-
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-
-@RunWith(AndroidJUnit4.class)
-public class UserCreation4Test {
-
-    @Rule
-    public ActivityScenarioRule<UserCreationPage4> testRule =
-            new ActivityScenarioRule<>(UserCreationPage4.class);
-
-
-    @Test
-    public void checkMailPage() {
-        int nextButtonMail = R.id.nextButtonMail;
-        onView(withId(nextButtonMail)).check(matches(isDisplayed()));
-        //onView(withId(nextButtonMail)).perform(click());
-    }
-
-
-
-}
